refactoring code structure and names
